feat: move websoc logic to backend
