+++ packages/types/src/index.ts
@@ -4,3 +4,4 @@ export * from './user';
 export * from './legacy';
 export * from './websoc';
 export * from '@packages/anteater-api-schemas/src/courses';
+export * from '@packages/anteater-api-schemas/src/websoc';
